Possible ad-hoc project creation for daily time sheet
This is a change request. When an employee is entering some daily time sheet, he may be entering some data for the project and customer that is not at the moment in the database. So, he must wait for the admin to create the data.

This way, we can allow him to enter the project name, we create a project; in this case we allow the project to have no customer for which I removed @NotNull constraint of the customer in the project.

Some minor other changes that may not be important.

Other thing that is on my mind is that how we can add custom error messages in this process. I tried to use some validator but I could not.