BUG: Spring MVC form hidden field appears null only in IE
To resolve this, we store the open timesheet in session so that it could be resolved.
This is DANGEROUS since the user should not work with the current month timesheet while working with an open timesheet until she closes it.

A question is opened at http://stackoverflow.com/questions/3676242/spring-mvc-form-hidden-field-appears-null-only-in-ie
when we have some update, we will resolve the bug.